a minimal refactoring of the entrypoint
